requirements.txt
pandas
numpy
scikit-learn
streamlit
gradio
matplotlib
seaborn
imdbpy
